pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chessai"
version = "0.1.0"
description = "Deep learning chess engine with MCTS and alpha-beta search"
authors = [{name = "ChessAI Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tensorflow>=2.15,<2.17",
    "tensorflow-addons==0.23.0",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "scikit-learn==1.4.2",
    "python-chess==1.999",
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "typer==0.12.3",
    "pyyaml==6.0.1",
    "orjson==3.10.6",
    "tensorboard==2.16.2",
    "matplotlib==3.8.4",
    "rich==13.7.1"
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88